Add NMAgent client
Many interactions with Azure SDN components occur through NMAgent, yet
there isn't a package for working expressly with NMAgent. This adds a
client as well as a transport for working with Wireserver, the proxy
used to communicate with NMAgent.

Other utilities are also present in this package for doing various
things like retries and validation of struct fields. These were useful
for the development of the client and could potentially be extracted
into their own packages in the future.